refactor(comp): cleanup - better structure code
Refs: #38
